remove 2 actions and 2 connectors
