+++ shared/actions/typed-actions-gen.tsx
@@ -244,8 +244,6 @@ export type TypedActionsMap = {
   'chat2:setBotSettings': chat2.SetBotSettingsPayload
   'chat2:removeBotMember': chat2.RemoveBotMemberPayload
   'chat2:editBotSettings': chat2.EditBotSettingsPayload
-  'chat2:loadMoreSmalls': chat2.LoadMoreSmallsPayload
-  'chat2:resetSmalls': chat2.ResetSmallsPayload
   'chat2:setParticipants': chat2.SetParticipantsPayload
   'chat2:findGeneralConvIDFromTeamID': chat2.FindGeneralConvIDFromTeamIDPayload
   'chat2:setGeneralConvFromTeamID': chat2.SetGeneralConvFromTeamIDPayload
