+++ shared/chat/inbox/index.desktop.tsx
@@ -4,9 +4,9 @@ import * as Styles from '../../styles'
 import * as T from './index.d'
 import * as Types from '../../constants/types/chat2'
 import AutoSizer from 'react-virtualized-auto-sizer'
-import BigTeamsDivider from './row/big-teams-divider/container'
+import BigTeamsDivider from './row/big-teams-divider'
 import BuildTeam from './row/build-team'
-import TeamsDivider from './row/teams-divider/container'
+import TeamsDivider from './row/teams-divider'
 import UnreadShortcut from './unread-shortcut'
 import * as Kb from '../../common-adapters'
 import {VariableSizeList} from 'react-window'
