+++ shared/chat/inbox/index.native.tsx
@@ -4,11 +4,11 @@ import * as RowSizes from './row/sizes'
 import * as Styles from '../../styles'
 import * as T from './index.d'
 import * as Types from '../../constants/types/chat2'
-import BigTeamsDivider from './row/big-teams-divider/container'
+import BigTeamsDivider from './row/big-teams-divider'
 import BuildTeam from './row/build-team'
 import ChatInboxHeader from './header/container'
 import InboxSearch from '../inbox-search/container'
-import TeamsDivider from './row/teams-divider/container'
+import TeamsDivider from './row/teams-divider'
 import UnreadShortcut from './unread-shortcut'
 import debounce from 'lodash/debounce'
 import {makeRow} from './row'
